Significant refactor of UI code to utilise tailwindcss 3 has been made to the `/` (home) page.
