+++ resources/css/tailwind.css
@@ -1,0 +1,4 @@
+/* tailwind */
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
