+++ resources/js/components/account-account-type-toggling-selector.vue
@@ -1,210 +1,180 @@
 <template>
-    <div class="field is-horizontal" v-bind:id="getIdForComponent">
-        <!--Account/Account-type-->
-        <div class="field-label is-normal no-padding-top">
-            <div class="label">
-                <toggle-button
-                    v-bind:id="getIdForToggleSwitch"
-                    v-model="propToggle"
-                    v-bind:value="propToggle"
-                    v-bind:labels="toggleButtonProperties.label"
-                    v-bind:color="toggleButtonProperties.colors"
-                    v-bind:height="toggleButtonProperties.height"
-                    v-bind:width="toggleButtonProperties.width"
-                    v-bind:sync="true"
-                />
-            </div>
-            <!--Account/Account-type display disabled checkbox -->
-            <div class="show-disabled-accounts-or-account-types" v-show="areDisabledAccountsOrAccountTypesPresent">
-                <input class="is-checkradio is-circle is-small" type="checkbox"
-                   v-bind:id="getIdForShowDisabledCheckbox"
-                   v-model="canShowDisabledAccountAndAccountTypes"
-                />
-                <label v-bind:for="getIdForShowDisabledCheckbox">Show Disabled</label>
-            </div>
-        </div>
-        <div class="field-body"><div class="field">
-            <div class="control">
-                <!--Account/Account-type selector -->
-                <div class="select" v-bind:class="{'is-loading': !areAccountsAndAccountTypesSet}">
-                    <select name="select-account-or-account-types-id" class="has-text-grey-dark select-account-or-account-types-id"
-                        v-model="propSelect"
-                        v-on:change="selectorValueChange"
-                    >
-                        <option value="" selected>[ ALL ]</option>
-                        <option
-                            v-for="accountOrAccountType in listProcessedAccountOrAccountTypes"
-                            v-show="!accountOrAccountType.disabled || canShowDisabledAccountAndAccountTypes"
-                            v-bind:key="accountOrAccountType.id"
-                            v-bind:value="accountOrAccountType.id"
-                            v-bind:class="{'disabled-option has-text-grey-light' : accountOrAccountType.disabled}"
-                            v-text="accountOrAccountType.name"
-                        ></option>
-                    </select>
-                </div>
-            </div>
-        </div></div>
+    <!--Account/Account-type-->
+  <div v-bind:id="getIdForComponent" class="flex justify-between">
+    <div class="flex flex-col mr-2">
+      <!-- Account/Account-type toggle button -->
+      <toggle-button
+          class="account-or-account-type-toggle-button"
+          v-bind="toggleButtonProperties"
+          v-bind:toggle-state.sync="accountOrAccountTypeToggledFromProps"
+      ></toggle-button>
+
+      <!-- Account/Account-type display disabled checkbox -->
+      <label class="show-disabled-accounts-or-account-types rounded-md py-0.5 px-2 mt-px inline-flex text-gray-700 text-xs"
+          v-show="areDisabledAccountsOrAccountTypesPresent"
+          v-bind:for="getIdForShowDisabledCheckbox"
+      >
+        <input type="checkbox" class="appearance-none hidden"
+            v-bind:id="getIdForShowDisabledCheckbox"
+            v-model="canShowDisabledAccountAndAccountTypes"
+        />
+        <svg xmlns="http://www.w3.org/2000/svg" class="h-4 w-4 -mt-px mr-1" fill="none" viewBox="0 0 24 24" stroke="currentColor"
+           v-show="!canShowDisabledAccountAndAccountTypes"
+        >
+          <circle cx="50%" cy="50%" r="10" stroke-width="2"  />
+        </svg>
+        <svg xmlns="http://www.w3.org/2000/svg" class="h-4 w-4 -mt-px mr-1 text-green-600 opacity-90 hover:opacity-100" fill="none" viewBox="0 0 24 24" stroke="currentColor"
+           v-show="canShowDisabledAccountAndAccountTypes"
+        >
+          <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z" />
+        </svg>
+        Show Disabled
+      </label>
     </div>
+
+    <!--Account/Account-type selector -->
+    <div class="flex w-full">
+      <div class="relative text-gray-700 w-full">
+        <span class="loading absolute inset-y-3 left-2 w-full" v-show="!areAccountsAndAccountTypesSet">
+          <svg class="animate-spin mr-3 h-5 w-5 text-blue-500" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
+            <circle class="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" stroke-width="4"></circle>
+            <path class="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
+          </svg>
+        </span>
+
+        <select id="entry-account-type" name="select-account-or-account-types-id" class="select-account-or-account-types-id rounded w-full"
+            v-model="accountOrAccountTypeIdFromProps"
+            v-on:change="selectorValueChange"
+        >
+          <option value="" selected>[ ALL ]</option>
+          <option
+              v-for="accountOrAccountType in listProcessedAccountOrAccountTypes"
+              v-show="!accountOrAccountType.disabled || canShowDisabledAccountAndAccountTypes"
+              v-bind:key="accountOrAccountType.id"
+              v-bind:value="accountOrAccountType.id"
+              v-text="accountOrAccountType.name"
+              v-bind:class="{'disabled-option has-text-grey-light' : accountOrAccountType.disabled}"
+          ></option>
+        </select>
+      </div>
+    </div>
+  </div>
 </template>
 
-<script>
-    import _ from "lodash";
-    import {accountsObjectMixin} from "../mixins/accounts-object-mixin";
-    import {accountTypesObjectMixin} from "../mixins/account-types-object-mixin";
-    import {bulmaColorsMixin} from "../mixins/bulma-colors-mixin";
-    import {ToggleButton} from 'vue-js-toggle-button'
+<script lang="js">
+import _ from "lodash";
+import {accountsObjectMixin} from "../mixins/accounts-object-mixin";
+import {accountTypesObjectMixin} from "../mixins/account-types-object-mixin";
+import {bulmaColorsMixin} from "../mixins/bulma-colors-mixin";
+import ToggleButton from "./toggle-button";
 
-    export default {
-        name: "account-account-type-toggling-selector",
-        components: {
-            ToggleButton
-        },
-        mixins: [accountsObjectMixin, accountTypesObjectMixin, bulmaColorsMixin],
-        props: {
-            id: {type: String, required: true},
-            accountOrAccountTypeToggled: {type: Boolean, default: true},
-            accountOrAccountTypeId: {type: String|Number, required: true, default: ''}
-        },
-        data: function(){
-            return {
-                selectedToggleSwitch: {
-                    account: true,
-                    accountType: false
-                },
+const EMIT_UPDATE_TOGGLE = 'update:accountOrAccountTypeToggled';
+const EMIT_UPDATE_SELECT = 'update:accountOrAccountTypeId';
 
-                propToggle: this.accountOrAccountTypeToggled,
-                propSelect: this.accountOrAccountTypeId,
-
-                canShowDisabledAccountAndAccountTypes: false,
-            }
-        },
-        watch: {
-            accountOrAccountTypeToggled: function(newValue, oldValue){
-                this.propToggle = newValue;
-            },
-            accountOrAccountTypeId: function(newValue, oldValue){
-                this.propSelect = newValue;
-            },
-            propToggle: function(newValue, oldValue){
-                this.resetAccountOrAccountTypeSelectValue();
-                this.$emit('update-toggle', newValue);
-            },
-        },
-        computed: {
-            listProcessedAccountOrAccountTypes: function(){
-                return this.processListOfObjects(this.listAccountOrAccountTypes, this.canShowDisabledAccountAndAccountTypes);
-            },
-            listAccountOrAccountTypes: function(){
-                if(this.propToggle === this.selectedToggleSwitch.account){
-                    return this.rawAccountsData;
-                } else if(this.propToggle === this.selectedToggleSwitch.accountType){
-                    return this.rawAccountTypesData;
-                }
-            },
-
-            areAccountsAndAccountTypesSet: function(){
-                return this.areAccountTypesAvailable && this.areAccountsAvailable;
-            },
-            areDisabledAccountsOrAccountTypesPresent: function(){
-                return !_.isEmpty(this.listAccountOrAccountTypes)
-                    && this.listAccountOrAccountTypes
-                        .filter(function(accountOrAccountTypeObject){
-                            return accountOrAccountTypeObject.disabled
-                        }).length > 0;
-            },
-
-            getIdForComponent: function(){
-                return 'account-or-account-type-toggling-selector-for-'+this.id;
-            },
-            getIdForShowDisabledCheckbox: function(){
-                return 'show-disabled-accounts-or-account-types-'+this.id+'-checkbox';
-            },
-            getIdForToggleSwitch: function(){
-                return 'toggle-account-and-account-types-for-'+this.id;
-            },
-
-            toggleButtonProperties: function(){
-                return {
-                    label: {checked: "Account", unchecked: "Account Type"},
-                    colors: {checked: this.colorGreyLight, unchecked: this.colorGreyLight},
-                    height: 36,
-                    width: 140,
-                };
-            },
-        },
-        methods: {
-            processListOfObjects: function(listOfObjects, canShowDisabled=true){
-                if(!canShowDisabled){
-                    listOfObjects = listOfObjects.filter(function(object){
-                        return !object.disabled;
-                    });
-                }
-                return _.orderBy(listOfObjects, 'name');
-            },
-            resetAccountOrAccountTypeSelectValue: function(){
-                this.propSelect = '';
-                this.selectorValueChange();
-            },
-            selectorValueChange: function(){
-                this.$emit('update-select', this.propSelect);
-            }
-        }
+export default {
+  name: "account-account-type-toggling-selector",
+  components: {
+    ToggleButton
+  },
+  mixins: [accountsObjectMixin, accountTypesObjectMixin, bulmaColorsMixin],
+  props: {
+    id: {type: String, required: true},
+    accountOrAccountTypeToggled: {type: Boolean, default: true},
+    accountOrAccountTypeId: {type: String|Number, required: true, default: ''}
+  },
+  data: function(){
+    return {
+      canShowDisabledAccountAndAccountTypes: false,
+      accountOrAccountTypeToggledFromProps: this.accountOrAccountTypeToggled, // toggle
+      accountOrAccountTypeIdFromProps: this.accountOrAccountTypeId,  // select
     }
+  },
+  watch: {
+    accountOrAccountTypeToggled: function(newValue, oldValue){
+      // update to prop.accountOrAccountTypeToggled (for toggle) from parent after init
+      this.accountOrAccountTypeToggledFromProps = newValue;
+    },
+    accountOrAccountTypeId: function(newValue, oldValue){
+      // update to prop.accountOrAccountTypeId (for select) from parent after init
+      this.accountOrAccountTypeIdFromProps = newValue;
+    },
+    accountOrAccountTypeToggledFromProps: function(newValue, oldValue){
+      // data.accountOrAccountTypeToggledFromProps updated
+      this.resetAccountOrAccountTypeSelectValue();
+      this.$emit(EMIT_UPDATE_TOGGLE, newValue);
+    },
+  },
+  computed: {
+    areAccountsAndAccountTypesSet: function(){
+      return this.areAccountTypesAvailable && this.areAccountsAvailable;
+    },
+    areDisabledAccountsOrAccountTypesPresent: function(){
+      return !_.isEmpty(this.listAccountOrAccountTypes)
+        && this.listAccountOrAccountTypes
+          .filter(function(accountOrAccountTypeObject){
+            return accountOrAccountTypeObject.disabled
+          }).length > 0;
+    },
+
+    getIdForComponent: function(){
+      return 'account-or-account-type-toggling-selector-for-'+this.id;
+    },
+    getIdForShowDisabledCheckbox: function(){
+      return 'show-disabled-accounts-or-account-types-'+this.id+'-checkbox';
+    },
+    getIdForToggleSwitch: function(){
+      return 'toggle-account-and-account-types-for-'+this.id;
+    },
+
+    listProcessedAccountOrAccountTypes: function(){
+        return this.processListOfObjects(this.listAccountOrAccountTypes, this.canShowDisabledAccountAndAccountTypes);
+    },
+    listAccountOrAccountTypes: function(){
+      if(this.accountOrAccountTypeToggledFromProps === this.selectedToggleSwitch.account){
+        return this.rawAccountsData;
+      } else if(this.accountOrAccountTypeToggledFromProps === this.selectedToggleSwitch.accountType){
+        return this.rawAccountTypesData;
+      }
+    },
+
+    toggleButtonProperties: function(){
+      return {
+        colorChecked: this.colorGreyLight,
+        colorUnchecked: this.colorGreyLight,
+        fontSize: 12, // px
+        height: 36,
+        labelChecked: "Account",
+        labelUnchecked: "Account Type",
+        toggleId: this.getIdForToggleSwitch,
+        width: 140,
+      };
+    },
+    selectedToggleSwitch: function(){
+      return {
+        account: true,
+        accountType: false
+      }
+    },
+  },
+  methods: {
+    processListOfObjects: function(listOfObjects, canShowDisabled=true){
+      if(!canShowDisabled){
+        listOfObjects = listOfObjects.filter(function(object){
+          return !object.disabled;
+        });
+      }
+      return _.orderBy(listOfObjects, 'name');
+    },
+    resetAccountOrAccountTypeSelectValue: function(){
+      this.accountOrAccountTypeIdFromProps = '';
+      this.selectorValueChange();
+    },
+    selectorValueChange: function(){
+      this.$emit(EMIT_UPDATE_SELECT, this.accountOrAccountTypeIdFromProps);
+    },
+  }
+}
 </script>
 
 <style lang="scss" scoped>
-    $font-size: 0.85rem;
-    $quarter-margin: 0.25rem;
-
-    .field-label{
-        &.is-normal{
-            font-size: $font-size;
-            margin-right: 0.75rem;
-            padding-top: 0.5rem;
-
-            &.no-padding-top{
-                padding-top: 0;
-            }
-        }
-
-        .label{
-            width: 10rem;
-        }
-    }
-
-    .vue-js-switch{
-        font-size: $font-size;
-    }
-    .select-account-or-account-types-id{
-        min-width: 15rem;
-        max-width: 19rem;
-    }
-
-    .show-disabled-accounts-or-account-types{
-        margin-top: -0.25rem;
-        margin-bottom: -0.5rem;
-
-        +label:after,
-        +label::after{
-            top: 0.3125rem;
-        }
-    }
-
-
-    // accounts/account-types toggle button
-    @import '~bulma/sass/helpers/color';
-    $toggle-button-bg-color: $grey-light;
-    $toggle-button-text-color: $white;
-    .account-or-account-type-toggle-button {
-        --toggle-width: 8.5rem;
-        --toggle-font-size: 0.8;
-        // accounts
-        --toggle-bg-on: #{$toggle-button-bg-color};
-        --toggle-border-on: #{$toggle-button-bg-color};
-        --toggle-text-on: #{$toggle-button-text-color};
-        // account-types
-        --toggle-bg-off: #{$toggle-button-bg-color};
-        --toggle-border-off: #{$toggle-button-bg-color};
-        --toggle-text-off: #{$toggle-button-text-color};
-    }
 </style>
