+++ resources/js/components/loading-modal.vue
@@ -1,60 +1,54 @@
 <template>
-    <div id="loading-modal" class="modal" v-bind:class="{'is-active' : isLoading}">
-        <div class="modal-background"></div>
-        <div class="modal-content">
-            <div class="container">
-                <spinner
-                    v-bind:depth="spinner.depth"
-                    v-bind:size="spinner.size"
-                    v-bind:speed="spinner.speed"
-                    v-bind:color="spinner.color"
-                ></spinner>
-            </div>
-        </div>
+  <div id="loading-modal" class="fixed inset-0 bg-gray-600 bg-opacity-50 overflow-y-auto h-full w-full" v-show="isLoading">
+    <div class="modal flex justify-center relative inset-y-1/4 mx-auto w-160">
+      <spinner v-bind="loaderProperties"></spinner>
     </div>
+  </div>
 </template>
 
 <script>
-    import randomColor from 'randomcolor/randomColor';
-    import Spinner from 'vue-spinner-component/src/Spinner.vue';
+import randomColor from 'randomcolor/randomColor';
+import Spinner from 'vue-spinner-component/src/Spinner.vue';
 
-    export default {
-        name: "loading-modal",
-        components: {
-            Spinner
-        },
-        data: function(){
-            return {
-                isLoading: false,
-                spinner: {
-                    depth: 5,
-                    size: 175,
-                    speed: 0.5,
-                    color: ""
-                },
-            };
-        },
-        methods: {
-            showLoading: function(){
-                this.generateRandomColor();
-                this.isLoading = true;
-            },
-            stopLoading: function(){
-                this.isLoading = false;
-            },
-            generateRandomColor: function(){
-                this.spinner.color = randomColor();
-            }
-        },
-        created: function(){
-            this.$eventHub.listen(this.$eventHub.EVENT_LOADING_SHOW, this.showLoading);
-            this.$eventHub.listen(this.$eventHub.EVENT_LOADING_HIDE, this.stopLoading);
-        },
+export default {
+  name: "loading-modal",
+  components: {
+    Spinner
+  },
+  data: function(){
+    return {
+      isLoading: false,
+      loadingColor: '',
+    };
+  },
+  computed: {
+    loaderProperties: function(){
+      return {
+        depth: 5,
+        size: 175,
+        speed: 0.5,
+        color: this.loadingColor
+      }
+    },
+  },
+  methods: {
+    showLoading: function(){
+      this.generateRandomColor();
+      this.isLoading = true;
+    },
+    stopLoading: function(){
+      this.isLoading = false;
+    },
+    generateRandomColor: function(){
+      this.loadingColor = randomColor();
     }
+  },
+  created: function(){
+    this.$eventHub.listen(this.$eventHub.EVENT_LOADING_SHOW, this.showLoading);
+    this.$eventHub.listen(this.$eventHub.EVENT_LOADING_HIDE, this.stopLoading);
+  },
+}
 </script>
 
 <style scoped>
-    .container {
-        width: 100px;
-    }
 </style>
