+++ resources/sass/tags-input.scss
@@ -1,7 +1,27 @@
 /* additional styling for voerro-tags-input component */
 .tags-input input{
   font-size: 0.75rem;
+  padding: 1px 2px;
+  line-height: normal;
 }
+
+.tags-input input:focus{
+  box-shadow: none;
+}
+
+.tags-input.active{
+  /* taken from the tailwindcss [type='text'] selecor */
+  outline: 2px solid transparent;
+  outline-offset: 2px;
+  --tw-ring-offset-width: 0px;
+  --tw-ring-offset-color: #fff;
+  --tw-ring-color: #2563eb;
+  --tw-ring-offset-shadow: var(--tw-ring-inset) 0 0 0 var(--tw-ring-offset-width) var(--tw-ring-offset-color);
+  --tw-ring-shadow: var(--tw-ring-inset) 0 0 0 calc(1px + var(--tw-ring-offset-width)) var(--tw-ring-color);
+  box-shadow: var(--tw-ring-offset-shadow), var(--tw-ring-shadow), var(--tw-shadow);
+  border-color: #2563eb;
+}
+
 .typeahead-badges{
   margin-top: 0.25rem;
 }
