+++ resources/views/head-component.blade.php
@@ -12,8 +12,7 @@
 <!-- Fonts -->
 <link href="https://fonts.googleapis.com/css?family=Raleway:100,600" rel="stylesheet" type="text/css">
 
-<link href="{{mix('vue/css/app.css')}}" rel="stylesheet" type="text/css">
-<link href="{{mix('vue/css/font-awesome.css')}}" rel="stylesheet" type="text/css">
+<link href="{{mix('dist/css/tailwind.css')}}" rel="stylesheet" type="text/css">
 
-<script type="text/javascript" src="{{mix('vue/js/manifest.js')}}"></script>
-<script type="text/javascript" src="{{mix('vue/js/vendor.js')}}"></script>
+<script type="text/javascript" src="{{mix('dist/js/manifest.js')}}"></script>
+<script type="text/javascript" src="{{mix('dist/js/vendor.js')}}"></script>
